+++ antifraud/src/main/java/com/bank/antifraud/service/SuspiciousPhoneTransfersServiceImp.java
@@ -8,6 +8,7 @@ import com.bank.antifraud.model.SuspiciousPhoneTransfers;
 import com.bank.antifraud.repository.SuspicionsPhoneTransfersRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -38,38 +39,44 @@ public class SuspiciousPhoneTransfersServiceImp implements SuspiciousPhoneTransf
 
     @Override
     @Transactional
-    public void savePhone(PhoneTransfersDTO newPhoneTransfers) {
-
+    public HttpStatus savePhone(PhoneTransfersDTO newPhoneTransfers) {
 
         log.info("Getting account transfer by id: {}", newPhoneTransfers);
 
         if (newPhoneTransfers.getPhone_transfer_id() == null) {
-            throw new ValidationException("Phone transfer ID must not be null.", "PhoneTransferServiceImpl.savePhone");
-        }
+            throw new ValidationException("Phone transfer ID must not be null.", "PhoneTransferServiceImpl.savePhone",
+                    HttpStatus.BAD_REQUEST);
 
+        }
         if (newPhoneTransfers.getSuspiciousReason().trim().length() < 3) {
             throw new ValidationException("Suspicious reason must not be null and must contain at least 3 characters",
-                    "SuspiciousPhoneTransfersServiceImp.savePhone");
+                    "PhoneTransferServiceImpl.savePhone", HttpStatus.BAD_REQUEST);
         }
 
         SuspiciousPhoneTransfers addPhone = phoneTransferMapper.toEntity(newPhoneTransfers);
         SuspiciousPhoneTransfers accountSave = sptRepository.save(addPhone);
         phoneTransferMapper.toDTO(accountSave);
+        return HttpStatus.OK;
     }
 
     @Override
     @Transactional
-    public void updatePhone(PhoneTransfersDTO updateTrDTO) {
+    public HttpStatus updatePhone(PhoneTransfersDTO updateTrDTO) {
 
         log.info("Creating phone transfer: {}", updateTrDTO);
 
         if (updateTrDTO.getPhone_transfer_id() == null) {
-            throw new ValidationException("Phone transfer ID must not be null.", "PhoneTransferServiceImpl.updatePhone");
+            throw new ValidationException(
+                    "Phone transfer ID must not be null.",
+                    "PhoneTransferServiceImpl.updatePhone",
+                    HttpStatus.BAD_REQUEST
+            );
         }
-
         if (updateTrDTO.getSuspiciousReason().trim().length() < 3) {
-            throw new ValidationException("Suspicious reason must not be null and must contain at least 3 characters",
-                    "PhoneTransferServiceImpl.updatePhone");
+            throw new ValidationException(
+                    "Suspicious reason must not be null and must contain at least 3 characters",
+                    "PhoneTransferServiceImpl.updatePhone",
+                    HttpStatus.BAD_REQUEST);
         }
 
         SuspiciousPhoneTransfers phoneTransfers = sptRepository.findById(updateTrDTO.getId())
@@ -79,20 +86,18 @@ public class SuspiciousPhoneTransfersServiceImp implements SuspiciousPhoneTransf
         phoneTransferMapper.updateEntityFromDTO(updateTrDTO, phoneTransfers);
         SuspiciousPhoneTransfers updateAccount = sptRepository.save(phoneTransfers);
         phoneTransferMapper.toDTO(updateAccount);
+        return HttpStatus.OK;
     }
 
     @Override
     public void delete(Long id) {
 
         log.info("Deleting phone transfer: {}", id);
-        try {
-            sptRepository.deleteById(id);
-        } catch (EntityNotFoundException ex) {
-
-            throw new EntityNotFoundException(
-                    "An object with this ID was not found for deletion, ID: " + id,
+        if (!sptRepository.existsById(id)) {
+            throw new EntityNotFoundException("An object with this ID was not found for deletion, ID: " + id,
                     "PhoneTransferServiceImpl.deletePhoneTransfer");
         }
+        sptRepository.deleteById(id);
     }
 
     @Override
